requirements.txt
flask>=2.0
gunicorn>=20.0
python-dotenv>=0.19
requests>=2.28
openai>=1.0
agentmail>=0.1
gspread>=5.0
google-auth>=2.0
